// front/src/containers/main/components/bonus/bonuses-types.ts
export enum LevelsActive {
    One = "/images/main/bonus/oneActive.png",
    Two = "/images/main/bonus/twoActive.png",
    Three = "/images/main/bonus/threeActive.png",
    Four = "/images/main/bonus/fourActive.png",
    Five = "/images/main/bonus/fiveActive.png",
}

export enum Levels {
    One = "/images/main/bonus/one.png",
    Two = "/images/main/bonus/two.png",
    Three = "/images/main/bonus/three.png",
    Four = "/images/main/bonus/four.png",
    Five = "/images/main/bonus/five.png",
}